# ICICI_Server/doc/swagger.json
{
    "info": {
        "version": "v0.0.1",
        "title": "ICICI",
        "description": "To perform any operation, you will need to first generate an access token. \n Tokens are generated by invoking the `/auth` API. \n The response is as follows, `Bearer <access-token>`. \n Copy the response and append it to all subsequent APIs where `Authorization` is required."
    },
...
    "paths": {
...
        "/api/project/": {
            "post": {
                "summary": "This endpoint will create project.",
                "description": "Create project.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created.",
                        "schema": {
                            "$ref": "#/responses/createProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/createProject"
                        },
                        "required": true,
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Project"
                ]
            },
            "get": {
                "summary": "This endpoint will get all the project details.",
                "description": "Project is an individual or collaborative enterprise and is the first and the utmost important component of the Application. This is where a procedure starts from. A Project interprets the name, version and instills a few actions. Name must be carefully planned to achieve a particular aim and it can be edited. There are actions like navigating to Products, navigating to Flows, editing the name & version, there is also a provision of Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. A version should be edited when its particular form which differs in respect to previous form. We can also afford to delete a project through actions. A Project can have multiple Products but only one Flow. This endpoint will create a project. Project is having project name, version and timestamp.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref": "#/responses/getProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Project"
                ]
            },
            "put": {
                "summary": "This endpoint will update project status.",
                "description": "Project is an individual or collaborative enterprise and is the first and the utmost important component of the Application. This is where a procedure starts from. A Project interprets the name, version and instills a few actions. Name must be carefully planned to achieve a particular aim and it can be edited. There are actions like navigating to Products, navigating to Flows, editing the name & version, there is also a provision of Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. A version should be edited when its particular form which differs in respect to previous form. We can also afford to delete a project through actions. A Project can have multiple Products but only one Flow. This endpoint will create a project. Project is having project name, version and timestamp.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref": "#/responses/createProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/updateProjectStatus"
                        },
                        "required": true,
                        "description": "No Description.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Project"
                ]
            },
            "delete": {
                "summary": "This endpoint will get the project details by projectId.",
                "description": "Project is an individual or collaborative enterprise and is the first and the utmost important component of the Application. This is where a procedure starts from. A Project interprets the name, version and instills a few actions. Name must be carefully planned to achieve a particular aim and it can be edited. There are actions like navigating to Products, navigating to Flows, editing the name & version, there is also a provision of Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. A version should be edited when its particular form which differs in respect to previous form. We can also afford to delete a project through actions. A Project can have multiple Products but only one Flow. This endpoint will update the project. Project is having project name, version and timestamp.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref": "#/responses/getProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Project"
                ]
            }
        },
        "/api/project/getAudit": {
            "get": {
                "summary": "This endpoint will get audit of the project.",
                "description": "A project is the second part of the Application. It is the catalogue of possessions of a Project. It is the sub-part of Project and can be navigated through Project. A Product like Projects engenders name, version and actions. Name is of the product and its version depicts the times its constructed. The actions include navigating to Flows, navigating to Services , arranges for Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. We can also delete a product through actions. A Project can assimilate only one flow and have multiple multiple Services.\nThis endpoint will get product by product Id.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/getProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "createdBy",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Project"
                ]
            }
        },
        "/api/project/initiateUAT": {
            "post": {
                "summary": "This endpoint will create project.",
                "description": "Create project.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created.",
                        "schema": {
                            "$ref": "#/responses/createProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/initiateUAT"
                        },
                        "required": true,
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Project"
                ]
            }
        },
        "/api/project/initiateSIT": {
            "post": {
                "summary": "This endpoint will initiate SIT.",
                "description": "initiate SIT.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created.",
                        "schema": {
                            "$ref": "#/responses/createProject"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/initiateSIT"
                        },
                        "required": true,
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Project"
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "description": "Authorization For Transaction APIs",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "createProject": {
            "type": "object",
            "required": [
                "productId",
                "photoUrls"
            ],
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "productId": {
                                "type": "string",
                                "description": ""
                            },
                            "services": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "serviceId": {
                                            "type": "string",
                                            "description": ""
                                        },
                                        "serviceName": {
                                            "type": "string",
                                            "description": ""
                                        }
                                    }
                                },
                                "description": "List of Terms and conditions for the disbursal.",
                                "format": "string"
                            }
                        }
                    },
                    "description": "List of Terms and conditions for the disbursal.",
                    "format": "string"
                },
                "productName": {
                    "type": "string",
                    "description": "Enter product name.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter org name.",
                    "format": "string"
                }
            }
        },
        "updateProjectStatus": {
            "type": "object",
            "required": [
                "productId",
                "photoUrls"
            ],
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "Enter projectId.",
                    "format": "string"
                },
                "status": {
                    "type": "string",
                    "description": "Enter status.",
                    "format": "string"
                },
                "event": {
                    "type": "string",
                    "description": "Enter event.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter org name.",
                    "format": "string"
                }
            }
        },
...
        "initiateUAT": {
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "Enter project id",
                    "format": "string"
                },
                "clientCode": {
                    "type": "string",
                    "description": "Enter client code",
                    "format": "string"
                },
                "productName": {
                    "type": "string",
                    "description": "Enter product name",
                    "format": "string"
                },
                "serviceName": {
                    "type": "string",
                    "description": "Enter service name",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter org name.",
                    "format": "string"
                }
            }
        },
        "initiateSIT": {
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "Enter project id",
                    "format": "string"
                },
                "clientCode": {
                    "type": "string",
                    "description": "Enter client code",
                    "format": "string"
                },
                "serviceName": {
                    "type": "string",
                    "description": "Enter service name",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter org name.",
                    "format": "string"
                }
            }
        },
...
    },
    "responses": {
...
        "createProject": {
            "required": [],
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "You will get project Id.",
                    "format": "string"
                },
                "projectName": {
                    "type": "string",
                    "description": "You will get project name.",
                    "format": "string"
                },
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
...
    },
...
    "security": [
        {
            "access_token": []
        }
    ],
    "swagger": "2.0"
}